requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.19.0
transformers>=4.30.0
accelerate>=0.21.0

# Image processing
Pillow>=9.5.0
opencv-python>=4.7.0

# Utilities
numpy>=1.24.0
tqdm>=4.65.0
requests>=2.31.0

# Optional but recommended for better performance
scipy>=1.10.0
ftfy>=6.1.1
spacy>=3.5.0

# Development tools
pytest>=7.3.1
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0